refactor: Rename DAF.h to VEQ.h and Add optimization command option to Makefile
