+++ include/VEQ.h
